refactor: login with useWallet
