+++ package.json
@@ -31,8 +31,8 @@
     ]
   },
   "dependencies": {
-    "@casperdash/usewallet": "^0.0.1",
-    "@casperdash/usewallet-core": "^0.0.1",
+    "@casperdash/usewallet": "^0.0.2",
+    "@casperdash/usewallet-core": "^0.0.2",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@hookform/resolvers": "^2.9.10",
