+++ src/enums/loginType.enum.ts
@@ -1,0 +1,4 @@
+export enum LoginTypeEnum {
+  ADMIN = 'admin',
+  USER = 'user',
+}
